create ami model for easier use
