+++ src/pytorch/ami_data/leye_neurons.txt
@@ -1,20 +1,20 @@
-conv1_1->
-conv1_2->
+relu1_1->42
+relu1_2->
 pool1->
-conv2_1->
-conv2_2->28,58,60,102
-pool2->28,102
-conv3_1->45,185
-conv3_2->126
-conv3_3->48,52,54,93,126,135,185,207,239,246
-pool3->173,181,192
-conv4_1->95,276,323,327,412
-conv4_2->15,328,362
-conv4_3->235,246,251,272,410,448,480
-pool4->25,46,75,148,171,187,251,329,355,403,471
-conv5_1->251,497,504
-conv5_2->
-conv5_3->79
-pool5->145,387
-fc6->
-fc7->2384
+relu2_1->
+relu2_2->28,102
+pool2->28,102,106
+relu3_1->9,150,181,199
+relu3_2->20,126
+relu3_3->1,114,173
+pool3->1,181
+relu4_1->33,192,209,259,267,323,378,410,456
+relu4_2->181,206,292,392,481
+relu4_3->25,45,75,98,148,171,181,187,245,251,346,355,468,480,493
+pool4->75,91,171,187,329,468,484
+relu5_1->4,43,50,56,73,157,236,296,306,317,388,468
+relu5_2->147,223,328,377
+relu5_3->387
+pool5->227
+relu6->
+relu7->
