+++ src/pytorch/ami_data/mouth_neurons.txt
@@ -1,20 +1,20 @@
-conv1_1->
-conv1_2->
+relu1_1->42
+relu1_2->
 pool1->
-conv2_1->
-conv2_2->28,60,102
+relu2_1->
+relu2_2->28,69,102
 pool2->69,102
-conv3_1->45,83,136,184,185,251
-conv3_2->45,126
-conv3_3->48,51,52,59,61,74,101,126,135,185,207,239
-pool3->12,51,59,117,132,144,151,173,182
-conv4_1->138,253,323,396
-conv4_2->15
-conv4_3->26,63,174,225,235,361
-pool4->0,15,58,82,104,105,138,149,213,235,238,273,329,330,388
-conv5_1->133,343
-conv5_2->130,267
-conv5_3->
-pool5->47,466
-fc6->
-fc7->
+relu3_1->116,117,150,181
+relu3_2->126,177,223
+relu3_3->11,32,40,48,51,59,98,109,117,122,132,135,144,151,173
+pool3->11,12,117,132,144,151,173
+relu4_1->26,42,94,116,164,179,232,246,253,271,308,312,323,412,440,446,466,470,496
+relu4_2->34,75,77,163,168,182,205,370,374,397,481,508
+relu4_3->0,33,58,104,105,167,235,238,339,388,442,446
+pool4->0,15,33,104,105,138,157,173,235,238,273
+relu5_1->24,81,133,143,198,272,402,467
+relu5_2->96,193
+relu5_3->233
+pool5->47,233
+relu6->
+relu7->
