+++ src/pytorch/ami_data/nose_neurons.txt
@@ -1,20 +1,20 @@
-conv1_1->
-conv1_2->
+relu1_1->42
+relu1_2->
 pool1->
-conv2_1->
-conv2_2->28,58,60,102
-pool2->28,102
-conv3_1->57,136
-conv3_2->
-conv3_3->48,101,135,174,185,200
-pool3->59
-conv4_1->327
-conv4_2->280,391,462
-conv4_3->386,472,495,496
-pool4->15,46,187,194,237,273,388,426,427,456,480
-conv5_1->1,72,177,251,386,453
-conv5_2->39,55,84,124,175,236,289,336
-conv5_3->
-pool5->273,353,370
-fc6->
-fc7->
+relu2_1->
+relu2_2->28
+pool2->28,102,106
+relu3_1->150,199
+relu3_2->
+relu3_3->174
+pool3->32,161,224
+relu4_1->33,204,246,271,278,312,323,410,419,446
+relu4_2->72,152,202,210,289,361,397,439
+relu4_3->15,33,46,98,108,170,187,194,236,256,326,331,346,373,388,417,480
+pool4->15,46,98,108,169,170,194,236,273,388,417,437,480
+relu5_1->69,88,317,442,483,484,509
+relu5_2->231,341
+relu5_3->273,353
+pool5->353
+relu6->
+relu7->
