+++ src/pytorch/ami_data/reye_neurons.txt
@@ -1,20 +1,20 @@
-conv1_1->
-conv1_2->
+relu1_1->42
+relu1_2->
 pool1->
-conv2_1->
-conv2_2->28,58,60,102
-pool2->28,102
-conv3_1->83,169,184,185,199
-conv3_2->
-conv3_3->4,48,101,135,185,207
-pool3->173,182
-conv4_1->
-conv4_2->184,448
-conv4_3->215,235,355
-pool4->0,15,17,91,171,238,258,274,329,446,475
-conv5_1->228,409
-conv5_2->104
-conv5_3->
+relu2_1->
+relu2_2->28,69,102
+pool2->28,102,106
+relu3_1->6,9,150,181
+relu3_2->126,141,223
+relu3_3->98,236
+pool3->11,98,236
+relu4_1->26,33,135,192,267,323,410
+relu4_2->74,263,392
+relu4_3->0,25,91,108,252,273,274,355,446,468
+pool4->25,91,108,171,274,329,355,468,475
+relu5_1->87,157,177,249,324,384,416,468,483
+relu5_2->232
+relu5_3->
 pool5->
-fc6->
-fc7->
+relu6->
+relu7->
