+++ src/pytorch/vgg_face_dag.py
@@ -6,54 +6,6 @@ Functions related to hooks are added by me
 """
 import torch
 import torch.nn as nn
-from collections import OrderedDict
-
-
-class Hook():
-    def __init__(self, name, module, weaken_param, strengthen_param0, strengthen_param1, ami_data=None, return_tensor=False):
-        self.hook = module.register_forward_hook(self.hook_fn)
-        self.name = name
-        self.output = None
-        self.ami_data = ami_data
-        self.weaken_neurons = None
-        self.strengthen_neurons = []
-        self.attri_neurons = []
-        self.processed = False
-        self.return_tensor = return_tensor
-        self.weaken_param = weaken_param
-        self.strengthen_param0 = strengthen_param0
-        self.strengthen_param1 = strengthen_param1
-
-    def hook_fn(self, module, input, output):
-        # self.input = input.detach().clone().cpu().numpy()
-        # output shape: batch x neurons x ...
-        if self.ami_data is not None:
-            if (not self.processed) and (not self.strengthen_neurons):
-                self.weaken_neurons = torch.ones(output.shape[1], device=output.device)
-                for n_l in self.ami_data:
-                    n_l = n_l[self.name]
-                    self.attri_neurons.extend(n_l)
-                    if not n_l: continue
-                    tmp_n = torch.zeros(output.shape[1], device=output.device)
-                    tmp_n[n_l] = 1
-                    self.strengthen_neurons.append(tmp_n)
-                    self.weaken_neurons *= (1-tmp_n)
-                self.processed = True
-            if not self.attri_neurons or not self.strengthen_neurons:
-                return output
-            new_output = neuron_AmI(output, self.weaken_neurons, self.strengthen_neurons,
-                                    self.attri_neurons, self.name, self.weaken_param,
-                                    self.strengthen_param0, self.strengthen_param1)
-            return new_output
-        else:
-            if self.return_tensor:
-                self.output = output.detach().clone()
-            else:
-                self.output = output.detach().clone().cpu().squeeze(0).numpy()
-
-    def close(self):
-        self.hook.remove()
-
 
 class Vgg_face_dag(nn.Module):
 
@@ -100,7 +52,6 @@ class Vgg_face_dag(nn.Module):
         self.relu7 = nn.ReLU(inplace=True)
         self.dropout7 = nn.Dropout(p=0.5)
         self.fc8 = nn.Linear(in_features=4096, out_features=2622, bias=True)
-        self.my_hooks = {}
 
     def forward(self, x0):
         x1 = self.conv1_1(x0)
@@ -143,73 +94,7 @@ class Vgg_face_dag(nn.Module):
         x37 = self.dropout7(x36)
         x38 = self.fc8(x37)
 
-        if len(self.my_hooks) > 0 and list(self.my_hooks.values())[0].ami_data is None:
-            res = OrderedDict()
-            for n, h in self.my_hooks.items():
-                if h.output is None:
-                    print(f'{n}.output is None')
-                res[n] = h.output
-            return res
-
         return x38
-
-    def set_ami_params(self, ami_weaken_parameter, ami_strengthen_parameter0, ami_strengthen_parameter1):
-        self.weaken_param = ami_weaken_parameter
-        self.strengthen_param0 = ami_strengthen_parameter0
-        self.strengthen_param1 = ami_strengthen_parameter1
-
-    def register_my_hook(self, skip_layers=[], ami_data=None, return_tensor=False):
-        self.my_hooks = OrderedDict()
-        if ami_data is None:
-            # set fake ami params to avoid errors
-            self.set_ami_params(0,0,0)
-        for name, module in self.named_modules():
-            if name not in skip_layers:
-                if len(list(module.children())) == 0:
-                    print(f'register hook for {name}')
-                    self.my_hooks[name] = Hook(name, module, self.weaken_param,
-                                               self.strengthen_param0, self.strengthen_param1,
-                                               ami_data, return_tensor)
-
-
-def neuron_AmI(x, weaken_neurons, strengthen_neurons, attri_neurons, layer_i, weaken_param, strengthen_param0, strengthen_param1):
-    x = x.squeeze(0)
-    if len(x.shape) == 1:
-        data = x
-    else:
-        data = torch.sum(x, (1,2))
-    attri_data = data[attri_neurons]
-    attri_mean = attri_data.mean()
-    attri_std = attri_data.std(unbiased=False)
-    attri_min = attri_data.min()
-
-    deviation = torch.zeros(x.shape[0], device=weaken_neurons.device)
-    if attri_std != 0:
-        deviation = torch.max(deviation, (data-attri_mean)/attri_std)
-    wkn = weaken(deviation, weaken_param) * weaken_neurons
-
-    stn = 1.
-    for st_n in strengthen_neurons:
-        deviation = torch.zeros(x.shape[0], device=weaken_neurons.device)
-        if attri_std != 0:
-            deviation = torch.abs(data-attri_min) / attri_std
-        t_stn = strengthen(deviation, strengthen_param0, strengthen_param1) * st_n + (1 - st_n)
-        stn *= t_stn
-    stn *= (1.-weaken_neurons)
-
-    stn += wkn
-    if len(x.shape) == 3:
-        stn.unsqueeze_(-1)
-        stn.unsqueeze_(-1)
-    return (x*stn).unsqueeze(0)
-
-
-def strengthen(x, strengthen_param0, strengthen_param1):
-    return strengthen_param0 - torch.exp(-x/strengthen_param1)
-
-
-def weaken(x, weaken_param):
-    return torch.exp(-x/weaken_param)
 
 
 def vgg_face_dag(weights_path=None, **kwargs):
